requirements.txt
dj-database-url==0.4.1
Django==1.9.7
django-cors-headers==1.1.0
django-extensions==1.6.7
django-rest-swagger==0.3.8
djangorestframework==3.3.3
gunicorn==19.6.0
PyYAML==3.11
six==1.10.0
whitenoise==3.2
